GMA_django/requirements.txt
absl-py==0.11.0
anyio==2.0.2
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.6.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==41.0.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==4.1.7
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
entrypoints==0.3
filelock==3.0.12
gast==0.4.0
grpcio==1.35.0
h5py==3.1.0
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyter-server==1.2.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lml==0.1.0
Markdown==3.3.3
MarkupSafe==1.1.1
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.5
oauthlib==3.2.2
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
pycparser==2.20
pyexcel==0.6.6
pyexcel-io==0.6.4
Pygments==2.7.4
PyJWT==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.1
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
texttable==1.6.3
tornado==6.1
traitlets==5.0.5
tzdata==2022.7
urllib3==1.26.3
virtualenv==20.4.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1